+++ nukkad_App/src/main/java/com/code/monks/nukkad/repositories/OrderRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface OrderRepository extends JpaRepository<OrderEntity,Long> {
+public interface OrderRepository extends JpaRepository<OrderEntity,Integer> {
 
    List<OrderEntity> findByStatusEnum(String status);
